Create ADC JOYSTICK TEST PROGRAM
This program is verified and it's main purpose is to test the joystick output voltage values to calibrate the sum and addition values of the main program. In the first test run, the X axis has an ADC neutral value of 127 (b'1000 0000') while the Y axis has an ADC neutral value of 132 (b '1000 0100').