ADC para Joystick
This programming is specific for a dual axis joystick. However, at this moment the configuration isn't fully complete. It has a working ADC where it takes the X axis values of the joystick which increases or decreases an output variable (the variable that will adjust the manual PWM) and sends it to PORTB. It's the X axis input is set in the RA0 pin. What is missing is to limit the output variable so that it starts in a middle value (default position for the servos) and then increase or decrease without overflowing.