ADC Update
Added the Y axis sum and subtraction for the Joystick as well as adjusting a sum and subtraction value which will eventually be the one that controls the manual PWM. On the X axis, it has been verified that the sum and subtraction don't overflow. However, the new problem is that when the joy stick is at it's fastest sum and subtraction it doesn't do either and the arithmetic operation is staled.